requirements.txt
django==4.2.7
django-admin==2.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
matplotlib==3.8.0
pillow==10.1.0
psycopg==3.1.12
psycopg2==2.9.9
psycopg2-binary==2.9.1
psycopg_c==3.1.12
django-environ==0.7.0
django-redis==5.0.0
redis==4.0.2
django-allauth==0.48.0
djoser==2.1.0
gunicorn==20.1.0
whitenoise==5.3.0
django-debug-toolbar==3.2.3
django-extensions==3.2.3
django-modeltranslation==0.17.3
selenium == 4.14.0
drf-spectacular == 0.26.5
openai == 1.3.6
uwsgi == 2.0.23
python-dotenv == 1.0.1